requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.0
gspread>=6.0.0
